Removed debugging statements. Branch navigator now displays on mouseover.
